+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 ### Fixed
 * [`no-deprecated`]: prevent false positive on commonjs import ([#3614][] @akulsr0)
 * [`no-unsafe`]: report on the method instead of the entire component (@ljharb)
+* [`no-deprecated`]: report on the destructured property instead of the entire variable declarator (@ljharb)
 
 [#3614]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3614